refactor(fault-injection): move HTTP/1.1 scenario/oracle to blackbull.fault_injection
Sprint 46 groups the existing client-side HTTP/1.1 scenario primitives
(`Scenario`, `Step`, `SendBytes`, `Sleep`, `ReadResponse`, `Abort`,
`ScenarioResult`) and the differential oracle (`Category`, `SideOutcome`,
`categorize`, `normalize_response`, `run_scenario`,
`ACCEPTED_CATEGORIES`, `PER_REQUEST_TIMEOUT_S`) under a new
`blackbull.fault_injection` namespace alongside the upcoming HTTP/2
fault-injection server.  The new namespace gives the launch story a
single anchor: client→server fault injection (HTTP/1.1) and server→client
fault injection (HTTP/2) live together.

- New modules: `blackbull.fault_injection.scenario_h1`,
  `blackbull.fault_injection.oracle_h1`, and the package `__init__`
  re-exporting both.
- Deprecation shims at `blackbull.client.scenario` and
  `blackbull.client.scenario_oracle` emit `DeprecationWarning` and
  forward to the new location.  Per the deprecation-window policy
  (3 releases AND 1 month, whichever later), removal no earlier than
  v0.45.0 / 2026-09-16.
- `blackbull.client` top-level re-exports of `Scenario` & friends keep
  working without a warning (they import directly from the new path),
  so back-compat for the common case is silent.
- Internal callers migrated: `blackbull/client/http1.py`,
  `tests/conformance/http1/test_http1_differential.py`,
  `tests/conformance/http1/fuzz/fuzz_http1.py`,
  `tests/conformance/http1/fuzz/make_seeds.py`.
- `oracle_h1.run_scenario` imports `HTTP1Client` lazily inside the
  function body to break the import cycle introduced by the move.
- All 990 unit tests pass; conformance tests still collect cleanly.
